+++ tests/test_resources_job.py
@@ -116,9 +116,10 @@ class LaunchTests(unittest.TestCase):
         runtime, that we do.
         """
         with client.test_mode as t:
+            # test with JSON job template extra_vars
             t.register_json('/job_templates/1/', {
                 'ask_variables_on_launch': True,
-                'extra_vars': 'spam: eggs',
+                'extra_vars': '{"spam": "eggs"}',
                 'id': 1,
                 'name': 'frobnicate',
                 'related': {'launch': '/job_templates/1/launch/'},
@@ -131,14 +132,34 @@ class LaunchTests(unittest.TestCase):
             with mock.patch.object(click, 'edit') as edit:
                 edit.return_value = '# Nothing.\nfoo: bar'
                 result = self.res.launch(1, no_input=False)
-                self.assertTrue(
-                    edit.mock_calls[0][1][0].endswith('spam: eggs'),
+                self.assertDictContainsSubset(
+                    {"spam": "eggs"},
+                    yaml.load(edit.mock_calls[0][1][0])
                 )
-            self.assertEqual(
-                json.loads(t.requests[3].body)['extra_vars'],
-                'foo: bar',
+            self.assertDictContainsSubset(
+                {'foo': 'bar'},
+                json.loads(json.loads(t.requests[3].body)['extra_vars'])
             )
             self.assertDictContainsSubset({'changed': True, 'id': 42}, result)
+
+            # test with YAML and comments
+            t.register_json('/job_templates/1/', {
+                'ask_variables_on_launch': True,
+                'extra_vars': 'spam: eggs\n# comment',
+                'id': 1,
+                'name': 'frobnicate',
+                'related': {'launch': '/job_templates/1/launch/'},
+            })
+            with mock.patch.object(click, 'edit') as edit:
+                edit.return_value = '# Nothing.\nfoo: bar'
+                result = self.res.launch(
+                    1, no_input=False
+                )
+                self.assertIn('# comment', edit.mock_calls[0][1][0])
+                self.assertDictContainsSubset(
+                    {"spam": "eggs"},
+                    yaml.load(edit.mock_calls[0][1][0])
+                )
 
     def test_job_template_variables(self):
         """Establish that job template extra_vars are combined with local
@@ -207,12 +228,13 @@ class LaunchTests(unittest.TestCase):
             with mock.patch.object(click, 'edit') as edit:
                 edit.return_value = '# Nothing.\nfoo: bar'
                 result = self.res.launch(1, no_input=False)
-                self.assertTrue(
-                    edit.mock_calls[0][1][0].endswith('spam: eggs'),
+                self.assertDictContainsSubset(
+                    {"spam": "eggs"},
+                    yaml.load(edit.mock_calls[0][1][0])
                 )
-            self.assertEqual(
-                json.loads(t.requests[2].body)['extra_vars'],
-                'foo: bar',
+            self.assertDictContainsSubset(
+                {'foo': 'bar'},
+                json.loads(json.loads(t.requests[2].body)['extra_vars'])
             )
             self.assertDictContainsSubset({'changed': True, 'id': 42}, result)
 
@@ -232,9 +254,9 @@ class LaunchTests(unittest.TestCase):
                             method='POST')
             result = self.res.launch(1, extra_vars=['foo: bar'])
 
-            self.assertEqual(
-                json.loads(t.requests[2].body)['extra_vars'],
-                'foo: bar',
+            self.assertDictContainsSubset(
+                {'foo': 'bar'},
+                json.loads(json.loads(t.requests[2].body)['extra_vars'])
             )
             self.assertDictContainsSubset({'changed': True, 'id': 42}, result)
 
@@ -254,9 +276,9 @@ class LaunchTests(unittest.TestCase):
                             method='POST')
             result = self.res.launch(1, extra_vars=['foo: bar'])
 
-            self.assertEqual(
-                json.loads(t.requests[2].body)['extra_vars'],
-                'foo: bar',
+            self.assertDictContainsSubset(
+                {'foo': 'bar'},
+                json.loads(json.loads(t.requests[2].body)['extra_vars'])
             )
             self.assertDictContainsSubset({'changed': True, 'id': 42}, result)